Improve type checking performance
Move from json-schema-to-typescript to zod-to-json-schema. The former, for even
moderate schemas, resulted in TypeScript errors "type instantiation is
excessively deep and possibly infinite". It seems cheaper to infer
the TypeScript shape from the Zod schema, and from that, the shape of parse
events, than to infer the TypeScript shape from a const JSON schema object.

As a result, we've removed the OpenAI adapter in favor of examples declaring a
Zod schema.